requirements.txt
beautifulsoup4==4.12.3
et-xmlfile==1.1.0
lxml==5.1.0
MouseInfo==0.1.3
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pillow==10.2.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.1
six==1.16.0
soupsieve==2.5
tzdata==2024.1
zipfile36==0.1.3
